古くて不要と思われるコメントアウトとcopycome関係を削除
